web/requirements.txt
flask
redis
pyyaml
requests
